move api related stuff to api slice
