+++ apps/admin/src/store/slices/api/constants.ts
@@ -1,0 +1,2 @@
+export const GATEWAY = import.meta.env.VITE_API_ENDPOINT as string;
+export const BACKEND_API = `${GATEWAY}/backend`;
