+++ apps/admin/src/store/slices/api/endpoints.ts
@@ -1,11 +1,7 @@
-export const GATEWAY = import.meta.env.VITE_API_ENDPOINT as string;
-export const BACKEND_API = `${GATEWAY}/backend`;
-
-export const apiAddWord = () => `${BACKEND_API}/add-word`;
-
-export const apiUnknownWords = (): string => `${BACKEND_API}/list`;
-export const apiApproveWords = () => `${BACKEND_API}/approve-words`;
-export const apiRejectWords = () => `${BACKEND_API}/reject-words`;
+export const apiAddWord = () => `/add-word`;
+export const apiUnknownWords = (): string => `/list`;
+export const apiApproveWords = () => `/approve-words`;
+export const apiRejectWords = () => `/reject-words`;
 export const apiDictionaryLanguages = (length = 5) => {
   const urlParams = new URLSearchParams({});
   if (length !== undefined) {
